test(rewrite): execution + pipeline tests for service / tableMethod root fetchers
GraphQLQueryTest (test-spec, executes against Sakila):
- queryTableMethod_popularFilms_returnsFilteredAndProjectsSelectedColumns:
  exercises the projection-wrapping path. Filters via SampleQueryService's
  WHERE clause and asserts framework projection picks only requested columns.
  Asserts QUERY_COUNT == 1 (single SELECT for the tableMethod).
- queryServiceTable_filmsByService_returnsRecordsThatFlowThroughColumnFetchers:
  exercises the service-as-source-of-truth path. Service runs its own SELECT;
  graphql-java's column fetchers walk the returned Result<FilmRecord>.
- queryServiceRecord_filmCount_returnsScalar: exercises the scalar-coercion
  path; service returns Integer, graphql-java coerces to GraphQL Int!.

GraphitronSchemaBuilderTest (pipeline-tier, asserts classifier behavior):
- SERVICE_AT_ROOT_WITH_CONNECTION_RETURN_REJECTED: structural FilmConnection
  return type → UnclassifiedField with the Invariants §1 message.
- TABLEMETHOD_AT_ROOT_WITH_CONNECTION_RETURN_REJECTED: same shape on the
  @tableMethod arm.
- SERVICE_AT_ROOT_WITH_SOURCES_PARAM_REJECTED: TestServiceStub.getWithSources
  takes a List<Row1<Integer>>, classified as ParamSource.Sources by
  ServiceCatalog. The classifier's §2 check rejects this at the @service arm
  with the documented message.

TestServiceStub gains getWithSources for the §2 negative case.

SampleQueryService.popularFilms: minRentalRate accepts Double rather than
Float — graphql-java's Float scalar coerces to Double Java-side, so a Float
parameter would receive a ClassCastException at runtime.

Totals: 710 unit tests + 134 test-spec tests (3 new GraphQLQueryTest +
3 new pipeline negative cases). All green.

+++ graphitron-rewrite/graphitron-rewrite-fixtures/src/main/java/no/sikt/graphitron/rewrite/test/services/SampleQueryService.java
@@ -35,9 +35,11 @@ public final class SampleQueryService {
      * to {@code Tables.FILM} per the @table-bound return type) and projects via
      * {@code FilmType.$fields(...)} over the returned {@code Table<FilmRecord>}.
      */
-    public static Table<FilmRecord> popularFilms(Table<FilmRecord> filmTable, Float minRentalRate) {
+    public static Table<FilmRecord> popularFilms(Table<FilmRecord> filmTable, Double minRentalRate) {
+        // graphql-java's GraphQL Float scalar coerces Java-side to Double, so signatures that take
+        // a Float would receive a ClassCastException; Double is the correct receive type.
         return Tables.FILM.where(Tables.FILM.RENTAL_RATE.greaterOrEqual(
-            org.jooq.impl.DSL.value(minRentalRate, Tables.FILM.RENTAL_RATE.getDataType())));
+            org.jooq.impl.DSL.value(minRentalRate.floatValue(), Tables.FILM.RENTAL_RATE.getDataType())));
     }
 
     /**
